Update resume.html
Converted html to be email ready.

fixed "<p>&nbsp;</p> " that was automatically inserting itself in the code when entered into tinymice, ctrlq.org, and other HTML email services. Believed to be caused when 2 separate comments in the code were placed one right after the other. 

                                          </tr>
                       <!-- End of Title -->
                       <!-- spacing -->
                                          <tr>
                                             <td 
                                          </tr>
                         <!-- End of spacing -->
                         <!-- content -->
                                          <tr>
                                             <td 
                                                

I deleted the extra comments as they are not currently needed.